fix: viewmodel 초기화 시에 initializer 추가
