+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -1,13 +1,13 @@
 package woowacourse.shopping
 
 import android.app.Application
-import woowacourse.shopping.data.DefaultCartRepository
-import woowacourse.shopping.data.DefaultProductRepository
+import woowacourse.shopping.data.CartRepository
+import woowacourse.shopping.data.ProductRepository
 import woowacourse.shopping.di.DIContainer
 
 class ShoppingApplication : Application() {
     init {
-        DIContainer.putInstance(DefaultProductRepository())
-        DIContainer.putInstance(DefaultCartRepository())
+        DIContainer.putInstance(ProductRepository())
+        DIContainer.putInstance(CartRepository())
     }
 }
