+++ app/src/main/java/woowacourse/shopping/data/CartRepository.kt
@@ -2,10 +2,19 @@ package woowacourse.shopping.data
 
 import woowacourse.shopping.model.Product
 
-interface CartRepository {
-    fun addCartProduct(product: Product)
+// TODO: Step2 - CartProductDao를 참조하도록 변경
+class CartRepository {
+    private val cartProducts: MutableList<Product> = mutableListOf()
 
-    fun getAllCartProducts(): List<Product>
+    fun addCartProduct(product: Product) {
+        cartProducts.add(product)
+    }
 
-    fun deleteCartProduct(id: Int)
+    fun getAllCartProducts(): List<Product> {
+        return cartProducts.toList()
+    }
+
+    fun deleteCartProduct(id: Int) {
+        cartProducts.removeAt(id)
+    }
 }
