+++ app/src/main/java/woowacourse/shopping/di/DIContainer.kt
@@ -1,22 +1,22 @@
 package woowacourse.shopping.di
 
+import kotlin.reflect.KClass
 import kotlin.reflect.full.primaryConstructor
-import kotlin.reflect.javaType
 
 object DIContainer {
-    val instances = mutableMapOf<Class<*>, Any>()
+    val instances = mutableMapOf<KClass<*>, Any>()
 
     inline fun <reified T : Any> putInstance(instance: T) {
-        instances[T::class.java] = instance
+        instances[T::class] = instance
     }
 
-    @OptIn(ExperimentalStdlibApi::class)
     inline fun <reified T : Any> inject(): T {
         val constructor = T::class.primaryConstructor ?: throw Exception()
         val parameters =
             constructor.parameters.map {
-                it.type.javaType
-            }.map { instances[it] }.toTypedArray()
+                val clazz = it.type.classifier as KClass<*>
+                instances[clazz]
+            }.toTypedArray()
         return constructor.call(*parameters)
     }
 }
