+++ app/src/main/java/woowacourse/shopping/ui/MainActivity.kt
@@ -6,6 +6,7 @@ import android.view.Menu
 import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityMainBinding
@@ -17,7 +18,9 @@ class MainActivity : AppCompatActivity() {
 
     private val viewModel: MainViewModel by viewModels {
         viewModelFactory {
-            injectViewModel<MainViewModel>()
+            initializer {
+                injectViewModel<MainViewModel>()
+            }
         }
     }
