+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
@@ -13,7 +14,9 @@ class CartActivity : AppCompatActivity() {
     private val binding by lazy { ActivityCartBinding.inflate(layoutInflater) }
     private val viewModel: CartViewModel by viewModels {
         viewModelFactory {
-            injectViewModel<CartViewModel>()
+            initializer {
+                injectViewModel<CartViewModel>()
+            }
         }
     }
     private lateinit var dateFormatter: DateFormatter
